Implemented MonkeyTyper::stream
Implemented the MonkeyTyper stream function. After checking if there are no other streams running for this particular MonkeyTyper, it loops calls for moveStream before pushing it's results to the channel. It only stops if the the prompt is repeated in the stream or if killStream is called. Additionally, it will pause when pause is called, and unpause when unpause is called.